requirements/test.txt
-r prod.txt

# For vulnerability check
safety==1.8.5

# For static analysis
pylint==2.4.1

# For testing
Flask-Testing==0.7.1